create redirectResolver to handle basepath condition
Express 2.x had a basepath config setting which was respected by the redirect functionality. This was essential for the app to work while proxied by www.sfu.ca with a non-root path.

This functionality was removed in Express 3. There are probably better ways to achieve the desired functionality (e.g. app mounting) but this seems to work fine.